day1/kura/requirements.txt
getmac==0.7.0
paho-mqtt==1.4.0
pkg-resources==0.0.0
